requirements.txt
python-dotenv
pyrofork
tgcrypto
motor
colorama
git+https://github.com/cleitonleonel/pyquotex.git
dnspython
flask
gunicorn
requests
setuptools
selenium
undetected-chromedriver
